requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.999
websockets==12.0
ollama==0.1.7
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1